backend/user-service/requirements.txt
# Component A dependencies
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
PyJWT
requests
python-dotenv
cryptography
